+++ poms/reports/builders/performance_item.py
@@ -36,7 +36,9 @@ class PerformancePeriod:
     def items(self):
         return self._items.values()
 
-    def get(self, portfolio: object = None, account: object = None, strategy1: object = None, strategy2: object = None, strategy3: object = None, create: object = None) -> object:
+    def get(self, portfolio: object = None, account: object = None, strategy1: object = None, strategy2: object = None,
+            strategy3: object = None,
+            create: object = None) -> object:
         if create is None:
             create = True
         key = PerformanceReportItem.make_item_key(